+++ src/components/formik/inputs/InputFormikControl.jsx
@@ -56,22 +56,25 @@ export const InputFormikControl = (props) => {
 
   return (
     <Field name={name}>
-      {({ field, form }) => (
-        <BaseControl
-          error={!!form.errors[field.name]}
-          required={required}
-          label={label}
-          helperText={form.errors[field.name]}
-          {...remainingProps}
-        >
-          <InputUnstyled
-            {...field}
-            type={type}
-            disabled={disabled}
-            components={{ Input: InputStyle }}
-          />
-        </BaseControl>
-      )}
+      {({ field, form }) => {
+        // console.log('field', field);
+        return (
+          <BaseControl
+            error={!!form.errors[field.name]}
+            required={required}
+            label={label}
+            helperText={form.errors[field.name]}
+            {...remainingProps}
+          >
+            <InputUnstyled
+              type={type}
+              disabled={disabled}
+              components={{ Input: InputStyle }}
+              {...field}
+            />
+          </BaseControl>
+        );
+      }}
     </Field>
   );
 };
