+++ src/hooks/useForm/useForm.js
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { setFormFields } from './setFormFields';
 import { setInitialValues } from './setInitialValues';
 import { setSchema } from './setSchema';
@@ -5,7 +6,10 @@ import { setSchema } from './setSchema';
 export const useForm = (dataHook, id) => {
   const dataHookResponse = dataHook(id);
 
-  const initialValues = setInitialValues(dataHookResponse, id);
+  const initialValues = useMemo(
+    () => setInitialValues(dataHookResponse, id),
+    [dataHookResponse, id],
+  );
 
   const validationSchema = setSchema(dataHookResponse);
